Refactor para reflejar mejor que la base de datos es una sola infraestructura, pero que cada caso de uso (feature) accede a ella a través de su propio adaptador, pidiendo sólo lo que precisa.
