+++ data/src/androidTest/java/com/franciscogarciagarzon/pricetracker/data/features/registerproduct/dao/StoreDaoTest.kt
@@ -5,6 +5,7 @@ import androidx.room.Room
 import androidx.test.core.app.ApplicationProvider
 import com.franciscogarciagarzon.pricetracker.data.database.AppDatabase
 import com.franciscogarciagarzon.pricetracker.data.database.Converters
+import com.franciscogarciagarzon.pricetracker.data.database.dao.StoreDao
 import com.franciscogarciagarzon.pricetracker.data.database.entity.StoreEntity
 import com.google.gson.Gson
 import kotlinx.coroutines.test.runTest
