+++ data/src/main/java/com/franciscogarciagarzon/pricetracker/data/di/RepositoryModule.kt
@@ -1,9 +1,9 @@
 package com.franciscogarciagarzon.pricetracker.data.di
 
 import com.franciscogarciagarzon.pricetracker.data.features.registerproduct.adapter.PurchaseRepositoryImpl
-import com.franciscogarciagarzon.pricetracker.data.features.registerproduct.dao.PriceRecordDao
-import com.franciscogarciagarzon.pricetracker.data.features.registerproduct.dao.ProductDao
-import com.franciscogarciagarzon.pricetracker.data.features.registerproduct.dao.StoreDao
+import com.franciscogarciagarzon.pricetracker.data.database.dao.PriceRecordDao
+import com.franciscogarciagarzon.pricetracker.data.database.dao.ProductDao
+import com.franciscogarciagarzon.pricetracker.data.database.dao.StoreDao
 import com.franciscogarciagarzon.pricetracker.domain.features.registerproduct.ports.outgoing.repositories.PurchaseRepository
 import dagger.Module
 import dagger.Provides
